% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyFunction.R
\name{shinyFunction}
\alias{shinyFunction}
\title{Shiny App function}
\usage{
shinyFunction(data = datasets::mtcars)

shinyFunction(data = datasets::mtcars)
}
\arguments{
\item{data}{A data frame to work with. Defaults to mtcars.}
}
\value{
Returns in interactive application that studies the dataset mtcars
}
\description{
Shiny app implementation as a function
}
\examples{
if (interactive()) {
  shinyFunction()
}
}
